+++ src/app/views/atlas/constants/colors.const.ts
@@ -1,0 +1,26 @@
+const BLUE_SHADES = [
+    '#001147',
+    '#002085',
+    '#002fc2',
+    '#003eff',
+    '#b8c9ff',
+];
+
+const EXCLUDED_COLOR = '#838996';
+
+const RED_SHADES = [
+    '#ffb8b8',
+    '#ff8f8f',
+    '#ff6666',
+    '#fc4949',
+    '#e60000',
+];
+
+const COLOR_PALETTE = [...BLUE_SHADES, ...RED_SHADES];
+
+export {
+    BLUE_SHADES,
+    COLOR_PALETTE,
+    EXCLUDED_COLOR,
+    RED_SHADES
+};